+++ common/Global.lua
@@ -1,0 +1,29 @@
+local addonName, MEM = ...
+
+local L = MEM.Localization
+
+-----------------------
+--- Global Funtions ---
+-----------------------
+
+function Memento_CompartmentOnEnter(self, button)
+	GameTooltip:ClearAllPoints()
+	GameTooltip:SetOwner(button, "ANCHOR_LEFT")
+
+	GameTooltip_SetTitle(GameTooltip, addonName)
+	GameTooltip_AddNormalLine(GameTooltip, MEM.ADDON_VERSION .. " (" .. MEM.ADDON_BUILD_DATE .. ")")
+	GameTooltip_AddBlankLineToTooltip(GameTooltip)
+	GameTooltip_AddHighlightLine(GameTooltip, L["minimap-button.tooltip"])
+
+	GameTooltip:Show()
+end
+
+function Memento_CompartmentOnLeave()
+    GameTooltip:Hide()
+end
+
+function HMemento_CompartmentOnClick(_, button)
+    if button == "RightButton" then
+        Settings.OpenToCategory(MEM.MAIN_CATEGORY_ID)
+    end
+end