+++ locale/localization.enUS.lua
@@ -1,4 +1,41 @@
-local L = LibStub("AceLocale-3.0"):NewLocale("Memento", "enUS", true)
+local _, MEM = ...
+
+MEM.Localization = setmetatable({},{__index=function(self,key)
+        geterrorhandler()("Memento (Debug): Missing entry for '" .. tostring(key) .. "'")
+        return key
+    end})
+
+local L = MEM.Localization
+
+-- Options
+
+L["options.general"] = "General Options"
+L["options.general.notification.name"] = "Chat Notification"
+L["options.general.notification.tooltip"] = "Activate or deactivate the notification in the chat after a combat."
+L["options.general.minimap-button.name"] = "Minimap Button"
+L["options.general.minimap-button.tooltip"] = "When this is enabled, the minimap button is displayed."
+
+L["options.screenshots"] = "screenshots"
+
+L["options.other"] = "Other Options"
+L["options.other.debug-mode.name"] = "Debug Mode"
+L["options.other.debug-mode.tooltip"] = "Enabling the debug mode displays additional information in the chat."
+
+L["options.about"] = "About"
+L["options.about.game-version"] = "Game Version"
+L["options.about.addon-version"] = "Addon Version"
+L["options.about.lib-version"] = "Library Version"
+L["options.about.author"] = "Author"
+
+L["options.about.button-github.name"] = "Feedback & Help"
+L["options.about.button-github.tooltip"] = "Opens a popup window with a link to GitHub."
+L["options.about.button-github.button"] = "GitHub"
+
+-- General
+
+L["minimap-button.tooltip"] = "|cnLINK_FONT_COLOR:Left-click|r to open the gold and currency overview.\n|cnLINK_FONT_COLOR:Right-click|r to open the options."
+
+--- OLD ENTRIES - CAN BE DELETED IN THE FUTURE
 
 local eventColor = "ffF2E699"
 local normalCoor = "ffFFD200"