+++ CHANGELOG.md
@@ -2,4 +2,11 @@
 ### [@project-version@](https://github.com/wow-addon-dev/Memento/tree/@project-version@) (@build-date@)
 [Full Changelog](@full-changelog@) - [Previous Releases](https://github.com/wow-addon-dev/Memento/releases)
 
-- Fixed: The options for the boss fight were no longer displayed [classic] [burning crusade - classic anniversary edition] [mist of pandaria - classic]
+**This is version 2.0 of Aurarium. With this release, the software code for all addons is being harmonized so that development will be easier in the future.**
+
+**Important note: All settings have been reset to their defaults.**
+
+- Added: Required dependency 'Arcane Wizard: Library'
+- Changed: License
+- Updated: Addon description
+- Refactoring of the addon structure and source code